Removed DayOfWeek, WeekOfMonth, DayOfMonth, TaskTargetSetRepresentation, and Task's dependency on TaskTargetSetView

+++ Mu/core-data/transformable/DayPattern.swift
@@ -19,12 +19,12 @@ class DayPattern: NSObject, NSSecureCoding {
         case dom = 3
     }
     
-    var daysOfWeek: [Int]
-    var weeksOfMonth: [Int]
-    var daysOfMonth: [Int]
+    var daysOfWeek: [Int16]
+    var weeksOfMonth: [Int16]
+    var daysOfMonth: [Int16]
     var type: patternType
     
-    init(dow: [Int], wom: [Int], dom: [Int]) {
+    init(dow: [Int16], wom: [Int16], dom: [Int16]) {
         self.daysOfWeek = dow
         self.weeksOfMonth = wom
         self.daysOfMonth = dom
@@ -42,9 +42,9 @@ class DayPattern: NSObject, NSSecureCoding {
     }
     
     required init(coder decoder: NSCoder) {
-        guard let dow = decoder.decodeObject(of: [DayPattern.self], forKey: "daysOfWeek") as? [Int],
-            let wom = decoder.decodeObject(of: [DayPattern.self], forKey: "weeksOfMonth") as? [Int],
-            let dom = decoder.decodeObject(of: [DayPattern.self], forKey: "daysOfMonth") as? [Int],
+        guard let dow = decoder.decodeObject(of: [DayPattern.self], forKey: "daysOfWeek") as? [Int16],
+            let wom = decoder.decodeObject(of: [DayPattern.self], forKey: "weeksOfMonth") as? [Int16],
+            let dom = decoder.decodeObject(of: [DayPattern.self], forKey: "daysOfMonth") as? [Int16],
             let typeValue = decoder.decodeObject(of: [DayPattern.self], forKey: "type") as? Int8
             else {
             exit(-1)
